+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.4] - 2020-02-02
+### Changed
+- `JSONKtorClient`: Switched read to use ByteArray instead of ByteBuffer
+- `JSONKtorClient`: Updated for changes to `json-stream` library
+
 ## [0.3.1] - 2020-01-31
 ### Changed
 - Minor changes to tests